requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bayesian-testing==0.8.0
beanie==1.29.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
expectation @ git+https://github.com/jakorostami/expectation@d6e05e1a215a294d118a55e5546e0bd73af3dbec
fastapi==0.115.6
h11==0.14.0
idna==3.10
Jinja2==3.1.5
lazy-model==0.2.0
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.6.1
numpy==1.26.4
pandas==2.2.3
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
rich==13.9.4
scipy==1.15.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
toml==0.10.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
